Increase binwidth for datalen from 1 to 4.
ggplot2, rendering to a bitmap, was missing bins because they were so
narrow (≈1500 bins in ≈500 pixels). It seemed to depend on how exactly
the bins were aligned relative to the pixel grid; if you resized the
output window in R, you could see bins come and go. Because the datalen
chart has many peaks that are 1 bin wide, it's easy to lose the most
important information. Bin in groups of 4 until we think of something
better.

I tried also rendering to a huge bitmap (4× scale) and then scaling
down, but that makes the labels way too small. Rendering to PDF and
converting to PNG mostly works, but in the syninterval diagrams there's
a bunch of crud in the body of the histogram where many narrow
rectangles were drawn.