Flip backward-labeled connections.
It appears that Bro got the direction of some connections wrong: it has
them with client port 443. From inspecition of the trace, it is clear
that the host with the port of 443 is really the server, as other
connections have the same IP with port 443 being connected to by a
variety of IP addresses on ephemeral ports.